backend/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
django-filter==24.3
django-environ==0.11.2
django-redis==5.4.0
django-celery-beat==2.7.0
django-extensions==3.2.3

# Database
psycopg2-binary==2.9.10

# Async / WebSocket
channels==4.2.0
channels-redis==4.2.1
daphne==4.1.2

# Task Queue
celery==5.4.0
redis==5.2.1

# Utilities
Pillow==11.0.0
python-dateutil==2.9.0
whitenoise==6.8.2

# Testing
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0
factory-boy==3.3.1

# Code Quality
black==24.10.0
isort==5.13.2
flake8==7.1.1

# Production
gunicorn==23.0.0
sentry-sdk==2.19.2
